Replace metadata suggests with depends
With Chef 11, a suggested cookbook it still added to the run list, but
its attributes and helpers are not loaded before it is processed. This
caused errors when the solr suggestion failed to correctly run the java
cookbook. Depending a cookbook ensures that its attributes are loaded
before it is parsed during run list compilation.